플러그인 파일 존재 여부 확인

+++ src/main/java/com/arushia/utilizevelocity/MainCommand.java
@@ -30,31 +30,32 @@ public class MainCommand implements SimpleCommand {
                     source.sendPlainMessage("Wrong usage!");
                     return;
                 }
-                for(int i = 1; i < args.length; i++){
-                    String pluginName = args[i];
-                    File pluginFile = null;
-                    for(File otherPluginFile : Objects.requireNonNull(new File(configData.serversFolder + "/" + configData.mainServerFolder + "/plugins/").listFiles())){
-                        if(!otherPluginFile.getName().endsWith(".jar")) continue;
-                        if(otherPluginFile.getName().toLowerCase().contains(pluginName.toLowerCase())){
-                            pluginFile = otherPluginFile;
-                            break;
-                        }
-                    }
-                    File pluginFolder = new File(configData.serversFolder + "/" + configData.mainServerFolder + "/plugins/" + pluginName);
 
-                    for(File otherServerFolder : Objects.requireNonNull(new File(configData.serversFolder).listFiles())){
-                        if(otherServerFolder.getName().equals(configData.mainServerFolder) || !otherServerFolder.isDirectory()) continue;
-                        UtilizeVelocity.getInstance().getLogger().info("Copied to {}", otherServerFolder.getName());
-                        try {
+                String pluginName = args[1];
+                File pluginFile = null;
+                for(File otherPluginFile : Objects.requireNonNull(new File(configData.serversFolder + "/" + configData.mainServerFolder + "/plugins/").listFiles())){
+                    if(!otherPluginFile.getName().endsWith(".jar")) continue;
+                    if(otherPluginFile.getName().toLowerCase().contains(pluginName.toLowerCase())){
+                        pluginFile = otherPluginFile;
+                        break;
+                    }
+                }
+                File pluginFolder = new File(configData.serversFolder + "/" + configData.mainServerFolder + "/plugins/" + pluginName);
+
+                for(File otherServerFolder : Objects.requireNonNull(new File(configData.serversFolder).listFiles())){
+                    if(otherServerFolder.getName().equals(configData.mainServerFolder) || !otherServerFolder.isDirectory()) continue;
+                    UtilizeVelocity.getInstance().getLogger().info("Copied to {}", otherServerFolder.getName());
+                    try {
+                        if(pluginFile != null) {
                             new File(otherServerFolder.getPath() + "/plugins/" + pluginName).deleteOnExit();
                             Files.copy(Objects.requireNonNull(pluginFile), new File(otherServerFolder.getPath() + "/plugins/" + pluginName + ".jar"));
-                            FileUtils.copyDirectory(pluginFolder, new File(otherServerFolder.getPath() + "/plugins/" + pluginName));
-                        } catch (IOException e) {
-                            throw new RuntimeException("Failed to copy plugin", e);
                         }
+                        FileUtils.copyDirectory(pluginFolder, new File(otherServerFolder.getPath() + "/plugins/" + pluginName));
+                    } catch (IOException e) {
+                        throw new RuntimeException("Failed to copy plugin", e);
                     }
-                    UtilizeVelocity.getInstance().getLogger().info("done!");
                 }
+                UtilizeVelocity.getInstance().getLogger().info("done!");
                 break;
         }
     }
